"Renamed ReturnInvoiceItemScreen to ReturnInvoiceDetailScreen, removed ReturnInvoiceItemScreen file, and updated ReturnInvoiceListScreen to navigate to ReturnInvoiceDetailScreen instead."
